+++ routers/__init__.py
@@ -1,3 +1,5 @@
+import asyncio
+import sys
 from fastapi.encoders import jsonable_encoder
 from fastapi.responses import JSONResponse
 
@@ -7,3 +9,18 @@ def error_handler(status_code, json_message):
         status_code=status_code,
         content=jsonable_encoder(json_message),
     )
+
+
+def asyncio_fix(func):
+    """Decorator for fixing asyncio bug on windows"""
+
+    def wrapper(*args):
+        if (
+            sys.version_info[0] == 3
+            and sys.version_info[1] >= 8
+            and sys.platform.startswith("win")
+        ):
+            asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
+        return func(*args)
+
+    return wrapper
