+++ scrapers/__init__.py
@@ -4,7 +4,7 @@ import asyncio
 from typing import List
 
 from pydantic import BaseModel
-from utils import asyncio_fix
+from routers import asyncio_fix
 
 
 HTTP_PROXY = os.environ.get("HTTP_PROXY", None)
@@ -12,21 +12,6 @@ HEADER_AIO = {
     "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36 Edg/114.0.1823.67",
     "Cookie": "fencekey=8e5j3p61b3k0a9b0e44c5bbcecafaa5a2",
 }
-
-
-def asyncio_fix(func):
-    """Decorator for fixing asyncio bug on windows"""
-
-    def wrapper(*args):
-        if (
-            sys.version_info[0] == 3
-            and sys.version_info[1] >= 8
-            and sys.platform.startswith("win")
-        ):
-            asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
-        return func(*args)
-
-    return wrapper
 
 
 class BaseScraper:
